Added 'generate_2d_scatterplot()' method that creates a visual representation of the results created by the SWOT network. The representation is consisted of 3 scatterplots and a histogram. Switched the old 'generate_scatterplots()' method to 'generate_3d_scatterplot' method. The 2d scatterplot method is now more neat and easier to use for a report.
All methods that produce graphics in the form of an image are now returning Base64 strings in order to be included in an html report.

Extended predict() method to calculate probabilities of the target FRC to be less than 0.2, 0.25 and 0.3 mg/L

Implemented 'generate_html_report(filename)' method which produces an html report and saves it on disk under the name 'filename'

Updated run_swot_script() to produce an html report and save it under a name specified as a command line argument.